mtd: rawnand: cafe: Remove the dma-off-on-id-read hack
Not only does the read-id instruction use the DMA channel, it doesn't
use the mmio data port either. This worked only because the ID happens
to stick in the data port after a reset for some reason.

In any case this is not necessary when we end up using the proper
"non-mem" data port.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>