[dropme] dt-bindings: pinctrl: pinctrl-single: Extend use of compatible strings
Currently users of pin controllers that use pinctrl-single don't follow
good device tree design practices in that they don't use compatible
strings for the actual hardware.

That is not a good idea because some controllers that are compatible
with pinctrl-single may provide extra functionality that's not supported
by pinctrl-single. If their device node lacks specific enough compatible
string it will be difficult to bind a more featureful driver to it
should such need arise in future.

The pin controller used on MMP2 and MMP3 is one such case: it can
happily set alternate functions and pin bias with pinctrl-single, but
there are also pins that control whether pin behavior on platform
suspend.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>